PR feedback: add types file
